+++ lib/src/main/java/com/github/rvesse/airline/help/Help.java
@@ -11,12 +11,18 @@ import com.github.rvesse.airline.help.cli.CliGlobalUsageSummaryGenerator;
 import com.github.rvesse.airline.model.CommandGroupMetadata;
 import com.github.rvesse.airline.model.CommandMetadata;
 import com.github.rvesse.airline.model.GlobalMetadata;
+import com.github.rvesse.airline.parser.AbbreviatedCommandFinder;
+import com.github.rvesse.airline.parser.AbbreviatedGroupFinder;
+import com.google.common.base.Predicate;
 
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.List;
 import java.util.concurrent.Callable;
 
+import static com.google.common.base.Predicates.compose;
+import static com.google.common.base.Predicates.equalTo;
+import static com.google.common.collect.Iterables.find;
 import static com.google.common.collect.Lists.newArrayList;
 
 @Command(name = "help", description = "Display help information")
@@ -62,42 +68,73 @@ public class Help implements Runnable, Callable<Void> {
 
         String name = commandNames.get(0);
 
-        // main program?
+        // Main program?
         if (name.equals(global.getName())) {
+            // Main program help
             new CliGlobalUsageGenerator().usage(global, out);
             return;
         }
 
-        // command in the default group?
-        for (CommandMetadata command : global.getDefaultGroupCommands()) {
-            if (name.equals(command.getName())) {
-                new CliCommandUsageGenerator().usage(global.getName(), null, command.getName(), command, out);
-                return;
-            }
+        // Predicates we may need
+        Predicate<? super CommandGroupMetadata> findGroupPredicate;
+        Predicate<? super CommandMetadata> findCommandPredicate;
+        //@formatter:off
+        findGroupPredicate = global.allowsAbbreviatedCommands() 
+                             ? new AbbreviatedGroupFinder(name, global.getCommandGroups()) 
+                             : compose(equalTo(name), CommandGroupMetadata.nameGetter());
+        //@formatter:on
+
+        // A command in the default group?
+        //@formatter:off
+        findCommandPredicate = global.allowsAbbreviatedCommands() 
+                               ? new AbbreviatedCommandFinder(name, global.getDefaultGroupCommands())
+                               : compose(equalTo(name), CommandMetadata.nameGetter());
+        //@formatter:on
+        CommandMetadata command = find(global.getDefaultGroupCommands(), findCommandPredicate, null);
+        if (command != null) {
+            // Command in default group help
+            new CliCommandUsageGenerator().usage(global.getName(), null, command.getName(), command, out);
+            return;
         }
 
-        // command in a group?
-        for (CommandGroupMetadata group : global.getCommandGroups()) {
-            if (name.endsWith(group.getName())) {
-                // general group help or specific command help?
-                if (commandNames.size() == 1) {
-                    new CliCommandGroupUsageGenerator().usage(global, group, out);
-                    return;
-                } else {
-                    String commandName = commandNames.get(1);
-                    for (CommandMetadata command : group.getCommands()) {
-                        if (commandName.equals(command.getName())) {
-                            new CliCommandUsageGenerator()
-                                    .usage(global.getName(), group.getName(), command.getName(), command, out);
+        // A command in a group?
+        CommandGroupMetadata group = find(global.getCommandGroups(), findGroupPredicate, null);
+        if (group != null) {
+            // General group help or specific group command help?
+            if (commandNames.size() == 1) {
+                // General group help
+                new CliCommandGroupUsageGenerator().usage(global, group, out);
+                return;
+            } else {
+                // Group command help
+                String commandName = commandNames.get(1);
+                //@formatter:off
+                findCommandPredicate = global.allowsAbbreviatedCommands() 
+                                       ? new AbbreviatedCommandFinder(commandName, group.getCommands())
+                                       : compose(equalTo(commandName), CommandMetadata.nameGetter());
+                //@formatter:on
+                command = find(global.getDefaultGroupCommands(), findCommandPredicate, null);
+                if (command != null) {
+                    new CliCommandUsageGenerator().usage(global.getName(), group.getName(), command.getName(), command,
+                            out);
 
-                            return;
-                        }
-                    }
+                    return;
+                }
+
+                // Didn't find an appropriate command
+                if (global.allowsAbbreviatedCommands()) {
+                    System.out.println("Unknown command " + name + " " + commandName + " or an ambiguous abbreviation");
+                } else {
                     System.out.println("Unknown command " + name + " " + commandName);
                 }
             }
         }
 
-        System.out.println("Unknown command " + name);
+        // Didn't find an appropriate group
+        if (global.allowsAbbreviatedCommands()) {
+            System.out.println("Unknown command " + name + " or an ambiguous abbreviation");
+        } else {
+            System.out.println("Unknown command " + name);
+        }
     }
 }